backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
distlib==0.3.9
distro==1.9.0
executing==2.2.0
fastapi==0.136.1
filelock==3.18.0
Flask==3.1.1
fonttools==4.53.1
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.14.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==3.0.2
matplotlib==3.9.1
matplotlib-inline==0.1.7
mock-open==1.4.0
mpmath==1.3.0
mysql-connector-python==9.4.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numpy==2.0.1
openai==2.36.0
opencv-python==4.11.0.86
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.13.4
pydantic_core==2.46.4
Pygments==2.19.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.1
regex==2025.7.34
requests==2.32.4
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
setuptools==80.9.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.5.0
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
ultralytics==8.3.161
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.47.0
virtualenv==20.31.2
wcwidth==0.2.13
Werkzeug==3.1.3
